fix : user to users & merge
