--- chat_nest/src/user/user.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('user')
-export class UserController {}
