--- chat_nest/src/user/user.module.ts
@@ -1,7 +1,0 @@
-import { Module } from '@nestjs/common';
-import { UserService } from './user.service';
-
-@Module({
-  providers: [UserService]
-})
-export class UserModule {}
