--- chat_nest/src/user/user.service.ts
@@ -1,19 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-@Injectable()
-export class UserService {
-  public validateEmail(email: string) {
-    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-    return emailRegex.test(email);
-  }
-
-  public validatePassword(password: string) {
-    const passwordRegex = /^(?=.*[a-zA-Z])(?=.*\d)[a-zA-Z\d@$!%*?&]{8,16}$/;
-    return passwordRegex.test(password);
-  }
-
-  public validateName(name: string) {
-    const koreanNameRegex = /^[가-힣]{2,6}$/;
-    return koreanNameRegex.test(name);
-  }
-}
