+++ vet_nest/src/users/__test__/users.controller.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { UsersController } from '../users.controller';
 
-describe('UsersController', () => {
+describe('UserController', () => {
   let controller: UsersController;
 
   beforeEach(async () => {
