+++ vet_nest/src/users/users.service.ts
@@ -1,4 +1,19 @@
 import { Injectable } from '@nestjs/common';
 
 @Injectable()
-export class UsersService {}
+export class UsersService {
+  public validateEmail(email: string) {
+    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+    return emailRegex.test(email);
+  }
+
+  public validatePassword(password: string) {
+    const passwordRegex = /^(?=.*[a-zA-Z])(?=.*\d)[a-zA-Z\d@$!%*?&]{8,16}$/;
+    return passwordRegex.test(password);
+  }
+
+  public validateName(name: string) {
+    const koreanNameRegex = /^[가-힣]{2,6}$/;
+    return koreanNameRegex.test(name);
+  }
+}
